Make navigation and menu sticky
Had to make a few changes to the way the navigation was being hidden/shown.
The transform/will-change on the body was breaking the `position: fixed` of
the navbar and menu.

Before, it was using transformX on the content and menu to make them slide left/right.
Now, it simply slides the menu over the content using the `left` property.

Hopefully this doesn't cause any undesired side effects. Tested it locally on desktop
and mobile sized browsers.